8th 2.2/requirements.txt.txt
PyQt5==5.15.9
pandas==2.1.0
fpdf==1.7.2
pdfplumber==0.10.2
PyMuPDF==1.23.5
reportlab==4.0.4
python-docx==1.1.0
ttkbootstrap==1.10.1
openpyxl==3.1.2